fix graphql extensions query
The extension manifest used to contain a "title" field, but that was phased out a while ago in favor of just having a name. The `src extensions list` command and other subcommands still tried to fetch the manifest title, which resulted in the following error:

```
GraphQL errors:
[
  {
    "locations": [
      {
        "column": 9,
        "line": 29
      }
    ],
    "message": "Cannot query field \"title\" on type \"ExtensionManifest\"."
  }
]
```

This fixes the GraphQL query to not use the (nonexistent) title field anymore.