Add a Setup section to the README and quick links
When guiding a Sourcegraph colleague through using `src` with Automation
the first blocker was that their `src` was talking to `sourcegraph.com`.

I realized that in this README.md here there was barely any mention of
`SRC_ENDPOINT`.

This PR tries to fix that by adding a separate `Setup` section and
adding quick links at the top of the README to the three most important
sections: Installation, Setup, Usage.