Only show "update campaign" message if it's supported
Quoting my comment here: I know that it's hacky to do this in the
formatting helper, but for now I think this is the easiest change to
make without introducing a whole lot of global state and/or mixing the
update property into the CampaignPlan.

If someone knows of a better way, let me know.