Take Docker image content digest into cache key
This fixes #112 by fixing a regression I introduced when taking over the
`src actions exec` functionality.

I lower-cased `imageContentDigest` so that users cannot set it. That had
the side effect of `src` itself not writing the `imageContentDigest`
to the cache key (because `json` needs the fields to be exported in
order to marshal/unmarshal them)

The fix here is to upper-case the field and add a separate validation
step to make sure that users don't set the cache field.