Log scopeQuery as it will be sent to Sourcegraph
This should help with the problem encountered in #125 by making it
easier to debug the scopeQuery.

The change here adds another logline to `src actions scope-query` when
the `-v` flag is given and changes the existing log in `src actions
exec` to show the scope query as it will be sent over the wire.