Add -create-plan flag to 'src action exec'
When the -create-plan flag is given the `src action exec` command will
create a campaign plan from the produced patches, making it not
necessary to pipe the output to `src campaign plan create-from-patches`.

If the execution of the action failed in a subset of the repositories,
the user is prompted and asked whether they would create a campaign plan
from the successfully produced patches or not.