Log action errors
Previously we only directed the user to the logfile, but there are some
errors that are independent of the action's step: failing to fetch the
ZIP archive, failing to extract ZIP archive, etc.

Those are not in the logfiles and as it turns out we never logged them.

This change fixes that.