git-serve: port src-expose serve to hidden sub command
This initial commit copy pastes the serve code from the Sourcegraph repo
at commit d0d94c90. It is a pretty much pure copy paste, just with the
addition of registring a src command and renaming src-expose to src
git-serve.

I'd like to continue developing this against master and when it is more
ready I'll document it in the CLI and other docs.