api: add Client.NewHTTPRequest
I want to make authenticated requests against the Sourcegraph API which
are not via GraphQL. In particular I am experimenting with Server Sent
Events endpoint. This commit adds NewHTTPRequest to the Client
interface, which allows access to an authenticated http.Request. This
will then be used to communicate with the non-GraphQL endpoint.