Allow setting temp dir used by campaigns through flag/env var
This addresses a customer report in which `src campaigns apply` couldn't
be run in CI because the docker-in-docker setup prohibits mounting
`/tmp` into containers.

It also fixes parts of
https://github.com/sourcegraph/src-cli/issues/316, since the temporary
directory is also where the logs are stored, but we might want to split
those locations up (in case a user wants to keep logs but not inside the
temporary directory).