src login command for guidance on configuring endpoint/token
You can run `src login https://sourcegraph.example.com` to verify that your `src` is authenticated to access that Sourcegraph instance. If not, it'll show you friendly steps to help you get it set up.

This will simplify all of our docs and in-product text that relies on users using the CLI. The steps to get them to the right state will be just:

1. Install the Sourcegraph CLI (`src`) at https://github.com/sourcegraph/src-cli [or link to releases]
1. Run `src login SOURCEGRAPH_URL` [and we can even fill in `SOURCEGRAPH_URL` in in-product text]

Also adds a deprecation warning if you have an old-style `src-config.json` file, and updates the README.md to highlight the `src login` command.