Fix regression and clean up log files again
While looking into how we store ZIP files I realised that log files are
not cleaned up. That's not a big issue, since `/tmp` is cleaned up by
the OS, but it's still a regression and not what we intended.

Digging deeper I ended up starting at the previous for much too long
before realising that the cleanup code was simply missing :)