Clean up shared code between `batch apply` and `preview`
This is a follow-up to #511 and #512.

It cleans up the code that's shared between `batch apply` and `previe`
by trying to make the boundaries between the "TUI +
execution"-glue-code, the `service` pkg, the `executor` pkg and the
helpers clearer.

A lot of this is subjective, of course (what does "clearer" even mean,
man), but I think less duplication is Good and less cross-talk between a
trio of components too.