Fix macOS panic by pulling images before determining workspace creator
I introduced this when re-organizing the code in `batch_common.go`, but
I didn't run into this because I've only developed on Linux since.

On macOS, though, `src` would panic when the
`DockerVolumeWorkspaceCreator` would check the `ImageUIDGUID` on each
step, because `step.image` was nil.

This fixes the issue by making sure that all images are pulled before
inspecting them.

As a requirement to the change it also pulls out the pulling of the
workspace creator image, which I think is actually an improvement
because it makes the code less decoupled.

Guess I should switch between macOS and Linux more often.