Set default log to src-cli.debug.log if build tag debug is used
I don't care about the specifics - filename, build tag, etc. - but I
*need* this. And the reason why I need it is that it's hard to log
something when our TUI is active. Not because of how the TUI is
implemented, but because the whole concept of an updating TUI doesn't
gel with printf-debugging.

And this is perfect for me: I can `tail -f` the log file and then use
`log.Printf` whereever I want.