Only checkout repo archive if steps will be executed
Since we use a "reference counting" mechanism when checking out archives
we shouldn't bump the ref count if we never execute steps.

Because only after executing the steps do we decrease the ref count and
thus "free" the archive to be cleaned up.

The problem is that if a `Task` has a cache hit then the ref will never
be decreased and the archive never cleaned up.