servegit: write to debug log git command errors
Previously we did http.Error. However, we have already set the HTTP
response code. Additionally we may have already written other data. So
we instead just write out the error message on a new line.

This error can often be swallowed by the git client. So we additionally
log the error to the debug log.