docker: remove description updates when publishing a release
We used to update the Docker Hub description for
https://hub.docker.com/r/sourcegraph/src-batch-change-volume-workspace
on release, but doing so reliably has proven to be difficult, as Docker
Hub doesn't provide a reliable API to do this. When it fails, the entire
release shows up as red, which is obviously not ideal!

Since this description rarely changes, and the pushes themselves are
reliable, this removes the description update and makes it a manual
process.