Add JSONLinesCache to write cache entries to stdout
This is part of https://github.com/sourcegraph/sourcegraph/issues/26929
and requires https://github.com/sourcegraph/sourcegraph/pull/27061 on
the `sourcegraph/sourcegraph` side.

It adds a `JSONLinesCache` that writes cache entries to stdout. It's
used when executing `src batch exec`, which results in cache entries
being transfered to the Sourcegraph server via the executor.