WIP of SSBC execution from the command line.
This is fragile, and probably sub-optimal in a bunch of ways right now.
The UI progress reporting is shoddy, the construction of the final
preview link is terrifying, and there's no separation of SSBC/CSBC
services.

But, there's one thing that is ready for us all to talk about: what noun
and verb do we use in the command structure? I've chosen `remote` and
`run`, respectively. `run` matches the UI. `remote` is there because
users won't really think of "server" or "SSBC" as concepts, but "remote"
feels like a good description of what's going on.

An open question: should the existing commands move to a `src batch
local` namespace, with everything aliased for now to make it work?

Another open question: should there even _be_ namespaces there?

Please discuss.