chore: remove bazel
Bazel was initially added here as a convenience for using this from the
monorepo, but ultimately, we don't use it at all. And as this repo is
merely a simple Go codebase with nothing fancy or slow tests, it's not
that a good case for Bazel. Finally, we're not using it in CI either,
it's GHA, so it has a tendency to drift.