chore(ci): always use go version in .tool-versions (#1198)
setup-go@v6 supports reading in .tool-versions which means we will now
always run the same version we run in local dev. This is also a better
version to use than go.mod since that is a minimum version.

Note: they have not tagged a release for this support, but we want to use
it. So for now we pin to the latest commit which supports it. Once there is
a tagged release I will update the workflow files.

Test Plan: CI